app/requirements.txt
pandas==1.4.1
mysql-connector-python==8.0.28
requests==2.27.1
python-dotenv==0.20.0
Flask==2.0.3
numpy==1.22.3
scikit-learn>=0.18
matplotlib>=1.4.3
streamlit>=1.0.0